Added Support to Lambda W
Changed the grammar to correctly validate lambda.

Changed the grammar to support Lambda as an option in the String w.

Changed the CYK to validate first if the w = to Lambda and if it is validate if is generated directly to avoid execute the algorithm.

Changed the global attribute to a boolean to facilitate the validations.

All made with Lau as inspiration.

Co-Authored-By: Laura Martínez <71099708+Laura-Martinez182@users.noreply.github.com>